requirements.txt
aiomcache==0.8.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dotenv==0.9.9
fastapi==0.129.2
fastapi-cache2==0.2.2
h11==0.16.0
idna==3.11
pendulum==3.2.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.2.0
requests==2.32.5
six==1.17.0
spotipy==2.25.2
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
